Start optimizing designs of conditional generation tasks
Currently, the code for MT and summarization are entangled. This PR tries to reasonably decouple them
