+++ .gitignore
@@ -33,3 +33,4 @@ log.res
 gen_samples.py
 explainaboard/tests/test_api.py
 profile_results
+demo.py
