+++ explainaboard/processors/__init__.py
@@ -4,6 +4,8 @@ from . import text_classification
 from . import named_entity_recognition
 from . import extractive_qa
 from . import conditional_generation
+from . import summarization
+from . import machine_translation
 
 from . import text_pair_classification
 from . import hellaswag
